Crate libra_json_rpc_client

Modules

errors

views

Structs

JsonRpcAsyncClient

JsonRpcBatch

JsonRpcClient

Enums

JsonRpcResponse

Traits

ResponseAsView
For JSON RPC views that are returned as part of a JsonRpcResponse instance, this trait can be used to extract the view from a JsonRpcResponse instance when applicable

Functions

get_response_from_batch

process_batch_response
Helper methods for basic payload processing ///