Trait libra_json_rpc_client::ResponseAsView

pub trait ResponseAsView: Sized {
    fn unexpected_response_error<T>(response: JsonRpcResponse) -> Result<T> { ... }

    fn from_response(_response: JsonRpcResponse) -> Result<Self> { ... }

    fn optional_from_response(
        _response: JsonRpcResponse
    ) -> Result<Option<Self>> { ... }

    fn vec_from_response(_response: JsonRpcResponse) -> Result<Vec<Self>> { ... }
}

For JSON RPC views that are returned as part of a JsonRpcResponse instance, this trait can be used to extract the view from a JsonRpcResponse instance when applicable

Provided methods

fn unexpected_response_error<T>(response: JsonRpcResponse) -> Result<T>

fn from_response(_response: JsonRpcResponse) -> Result<Self>

fn optional_from_response(_response: JsonRpcResponse) -> Result<Option<Self>>

fn vec_from_response(_response: JsonRpcResponse) -> Result<Vec<Self>>

Loading content...